+++ packages/in_app_purchase/in_app_purchase_storekit/example/ios/RunnerTests/TranslatorTests.m
@@ -11,6 +11,7 @@
 
 @property(strong, nonatomic) NSDictionary *periodMap;
 @property(strong, nonatomic) NSMutableDictionary *discountMap;
+@property(strong, nonatomic) NSMutableDictionary *discountMissingIdentifierMap;
 @property(strong, nonatomic) NSMutableDictionary *productMap;
 @property(strong, nonatomic) NSDictionary *productResponseMap;
 @property(strong, nonatomic) NSDictionary *paymentMap;
@@ -39,6 +40,15 @@
     self.discountMap[@"identifier"] = @"test offer id";
     self.discountMap[@"type"] = @(SKProductDiscountTypeIntroductory);
   }
+  self.discountMissingIdentifierMap = [[NSMutableDictionary alloc] initWithDictionary:@{
+    @"price" : @"1",
+    @"priceLocale" : [FIAObjectTranslator getMapFromNSLocale:NSLocale.systemLocale],
+    @"numberOfPeriods" : @1,
+    @"subscriptionPeriod" : self.periodMap,
+    @"paymentMode" : @1,
+    @"identifier" : [NSNull null],
+    @"type" : @0,
+  }];
 
   self.productMap = [[NSMutableDictionary alloc] initWithDictionary:@{
     @"price" : @"1",
@@ -226,6 +236,15 @@
       XCTAssertEqualObjects(
           error, @"When specifying a payment discount the 'identifier' field is mandatory.");
     }
+  }
+}
+
+- (void)testGetMapFromSKProductDiscountMissingIdentifier {
+  if (@available(iOS 12.2, *)) {
+    SKProductDiscountStub *discount =
+        [[SKProductDiscountStub alloc] initWithMap:self.discountMissingIdentifierMap];
+    NSDictionary *map = [FIAObjectTranslator getMapFromSKProductDiscount:discount];
+    XCTAssertEqualObjects(map, self.discountMissingIdentifierMap);
   }
 }