+++ packages/in_app_purchase/in_app_purchase_storekit/lib/src/store_kit_wrappers/sk_product_wrapper.g.dart
@@ -42,7 +42,7 @@ SKProductDiscountWrapper _$SKProductDiscountWrapperFromJson(Map json) =>
           (json['subscriptionPeriod'] as Map?)?.map(
         (k, e) => MapEntry(k as String, e),
       )),
-      identifier: json['identifier'] as String? ?? '',
+      identifier: json['identifier'] as String? ?? null,
       type:
           const SKProductDiscountTypeConverter().fromJson(json['type'] as int?),
     );